+++ cardscan/build.gradle
@@ -28,7 +28,8 @@ dependencies {
     implementation 'com.android.support:support-annotations:28.0.0'
     implementation "com.android.support:appcompat-v7:28.0.0"
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.getbouncer:base:1.0.4014'
+    //implementation 'com.getbouncer:base:1.0.4014'
+    implementation project(":base")
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
